# Optimize import performance using numpy