@ssbucarlos 2021-12-30
this is a a pretty nice speedup, ty for the change